stage_api/requirements.txt
fastapi
uvicorn[standard]
python-multipart
opencv-python-headless
rapidocr-onnxruntime
asyncpg
pydantic
httpx
numpy
Pillow
pytesseract
regex
PyNaCl
